requirements.txt
jupyter
langchain-community
langgraph
langchain-anthropic
tavily-python
docker
ollama
python-dotenv